c40-v55

+++ blockchain/TxOutput.go
@@ -2,7 +2,7 @@ package blockchain
 
 type TxOutput struct {
 	//交易输出的值
-	Value int
+	Value int64
 	//交易输出的所有者
 	ScriptPubKey string
 }
